requirements.txt
streamlit
psycopg2-binary
sqlite3  # (opcional, normalmente já embutido no Python)
pandas
bcrypt
python-dotenv
